+++ src/commands/index.ts
@@ -3,4 +3,5 @@
  */
 
 export { initCommand } from './init';
+export { scanCommand } from './scan';